Change int to size_t

+++ Team34/Code34/src/spa/src/pkb/cfg.cpp
@@ -11,14 +11,14 @@ namespace pkb
 {
     using StatementNum = simple::ast::StatementNum;
 
-    CFG::CFG(int v)
+    CFG::CFG(size_t v)
     {
         total_inst = v;
-        adj_mat = new int*[v];
-        for(int i = 0; i < v; i++)
+        adj_mat = new size_t*[v];
+        for(size_t i = 0; i < v; i++)
         {
-            this->adj_mat[i] = new int[v];
-            for(int j = 0; j < v; j++)
+            this->adj_mat[i] = new size_t[v];
+            for(size_t j = 0; j < v; j++)
             {
                 adj_mat[i][j] = INF;
             }
@@ -35,16 +35,16 @@ namespace pkb
     std::string CFG::getMatRep()
     {
         auto res = zpr::sprint("      ");
-        for(int i = 0; i < total_inst; i++)
+        for(size_t i = 0; i < total_inst; i++)
         {
             res += zpr::sprint("{03} ", i + 1); 
         }
         res += zpr::sprint("\n"); 
 
-        for(int i = 0; i < total_inst; i++)
+        for(size_t i = 0; i < total_inst; i++)
         {
             res += zpr::sprint("{03} | ", i+1);
-            for(int j = 0; j <total_inst; j++)
+            for(size_t j = 0; j < total_inst; j++)
             {
                 res += zpr::sprint("{03} ", adj_mat[i][j] == INF ? -1 : adj_mat[i][j]);
             }
@@ -56,11 +56,11 @@ namespace pkb
     void CFG::computeDistMat()
     {
         // Adapted from https://www.geeksforgeeks.org/floyd-warshall-algorithm-dp-16/
-        for(int k = 0; k < total_inst; k++)
+        for(size_t k = 0; k < total_inst; k++)
         {
-            for(int i = 0; i < total_inst; i++)
+            for(size_t i = 0; i < total_inst; i++)
             {
-                for(int j = 0; j < total_inst; j++)
+                for(size_t j = 0; j < total_inst; j++)
                 {
                     if(adj_mat[i][j] > (adj_mat[i][k] + adj_mat[k][j]) &&
                         (adj_mat[k][j] != INF && adj_mat[i][k] != INF))
